refactor(55-04): extract ipns-record.codec.ts from ipns.service.ts
- Create ipns-record.codec.ts with IpnsRecordFields interface + 3 standalone fns
- parseIpnsRecordBytes, parseCachedRecord, withCachedPublicKey take Logger as param
- Remove the 3 private methods from IpnsService; replace calls with imported fns
- IpnsService retains @Injectable, constructor, and all orchestration unchanged
- No api:generate (no DTO/endpoint change); 893 API tests pass
